restructure package Laboratory_Work_No_3
