+++ src/main/kotlin/Laboratory_Work_No_3/ArrayDownQueue.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task1
+package Laboratory_Work_No_3
 
 class ArrayDownQueue<T>(
     private val initialSize: Int
