+++ src/main/kotlin/Laboratory_Work_No_3/ArrayUpQueue.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task1
+package Laboratory_Work_No_3
 
 class ArrayUpQueue<T>(
     private val initialSize: Int
