+++ src/main/kotlin/Laboratory_Work_No_3/Dineable.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task2
+package Laboratory_Work_No_3
 
 interface Dineable {
     companion object {
