+++ src/main/kotlin/Laboratory_Work_No_3/ElectricStation.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task2
+package Laboratory_Work_No_3
 
 class ElectricStation : Refuelable {
     override fun refuel(carId: String) {
