+++ src/main/kotlin/Laboratory_Work_No_3/GasStation.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task2
+package Laboratory_Work_No_3
 
 class GasStation : Refuelable {
     override fun refuel(carId: String) {
