+++ src/main/kotlin/Laboratory_Work_No_3/Node.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task1
+package Laboratory_Work_No_3
 
 data class Node<T>(
     val data: T,
