+++ src/main/kotlin/Laboratory_Work_No_3/RobotDinner.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3.Task2
+package Laboratory_Work_No_3
 
 class RobotDinner : Dineable {
     override fun serveDinner(carId: String) {
