+++ src/main/kotlin/Laboratory_Work_No_3/main.kt
@@ -1,14 +1,5 @@
-package Laboratory_Work_No_3.Task5
+package Laboratory_Work_No_3
 
-import Laboratory_Work_No_3.Task1.ArrayDownQueue
-import Laboratory_Work_No_3.Task1.ArrayUpQueue
-import Laboratory_Work_No_3.Task1.LinkedListQueue
-import Laboratory_Work_No_3.Task2.ElectricStation
-import Laboratory_Work_No_3.Task2.GasStation
-import Laboratory_Work_No_3.Task2.PeopleDinner
-import Laboratory_Work_No_3.Task2.RobotDinner
-import Laboratory_Work_No_3.Task3.CarStation
-import Laboratory_Work_No_3.Task4.Semaphore
 import kotlinx.coroutines.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
