+++ src/test/kotlin/CarStationTests.kt
@@ -1,13 +1,13 @@
-import Laboratory_Work_No_3.Task1.ArrayDownQueue
-import Laboratory_Work_No_3.Task1.ArrayUpQueue
-import Laboratory_Work_No_3.Task1.LinkedListQueue
-import Laboratory_Work_No_3.Task1.Queue
-import Laboratory_Work_No_3.Task2.ElectricStation
-import Laboratory_Work_No_3.Task2.GasStation
-import Laboratory_Work_No_3.Task2.PeopleDinner
-import Laboratory_Work_No_3.Task2.RobotDinner
-import Laboratory_Work_No_3.Task3.Car
-import Laboratory_Work_No_3.Task3.CarStation
+import Laboratory_Work_No_3.ArrayDownQueue
+import Laboratory_Work_No_3.ArrayUpQueue
+import Laboratory_Work_No_3.LinkedListQueue
+import Laboratory_Work_No_3.Queue
+import Laboratory_Work_No_3.ElectricStation
+import Laboratory_Work_No_3.GasStation
+import Laboratory_Work_No_3.PeopleDinner
+import Laboratory_Work_No_3.RobotDinner
+import Laboratory_Work_No_3.Car
+import Laboratory_Work_No_3.CarStation
 import org.junit.jupiter.api.Test
 import kotlin.random.Random
 import kotlin.test.assertEquals
