+++ src/test/kotlin/TrackerTests.kt
@@ -1,4 +1,4 @@
-import Laboratory_Work_No_3.Task2.*
+import Laboratory_Work_No_3.*
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
